@@ -12,11 +12,11 @@
 		}
 	],
 	"require": {
-		"php": "^5.6 || ^7.0",
+		"php": "^7.1",
 		"guzzlehttp/guzzle": "^6.2"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^5.5",
+		"phpunit/phpunit": "^7.5",
 		"phpdocumentor/phpdocumentor": "^2.9",
 		"mikey179/vfsStream": "^1.2"
 	},
